requirements.txt
transformers
pytorch
matplotlib